JobPostService/requirements.txt
#
#Ensure pip and setuptools are up to date before install packages below!!
#
appdirs==1.4.3
coreapi==2.3.0
coreschema==0.0.4
Django==1.10.6
django-filter==1.0.1
djangorestframework==3.6.2
djangorestframework-camel-case==1.0b1
django-unixdatetimefield==0.1.6
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==1.0
MySQL-python==1.2.5
packaging==16.8
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
uritemplate==3.0.0
django-cors-headers==2.0.2
